Sunk ErrorReason into MockApnsServerHandler.
